pyproject.toml
[tool.poetry]
name = "ofertas-scrap"
version = "0.1.0"
description = ""
authors = ["Rhanyele Marinho <rhany.marinho@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.12.2"
confluent-kafka = "^2.4.0"
requests = "^2.32.2"
bs4 = "^0.0.2"
logging = "^0.4.9.6"
pydantic = "^2.7.1"
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
